+++ stickerfinder/models/chat.py
@@ -59,7 +59,9 @@ class Chat(base):
     last_sticker_message_id = Column(BigInteger)
 
     # ForeignKeys
-    current_task_id = Column(UUID(as_uuid=True), ForeignKey('task.id', ondelete='cascade'), index=True)
+    current_task_id = Column(UUID(as_uuid=True),
+                             ForeignKey('task.id', ondelete='cascade', name='chat_current_task_id_fkey'),
+                             index=True)
     current_sticker_file_id = Column(String, ForeignKey('sticker.file_id', ondelete='SET NULL'), index=True)
 
     # Relationships